pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[project]
name = "ylo-igbn"
version = "0.0.1"
requires-python = ">=3.12"
dependencies = [
    "torch==2.3.0+cu118",
    "torchaudio==2.3.0+cu118",
    "torchvision==0.18.0+cu118",
    "Jinja2==3.1.3",
    "notebook",
    "numpy",
    "pandas",
    "matplotlib",
    "scipy",
    "face_recognition",
    "opencv-python",
    "boto3",
    "azure-ai-documentintelligence==1.0.0b2",
    "PySide6",
    "python-dotenv",
    "Django~=5.0.3",
    "djangorestframework~=3.15.1",
    "drf-spectacular~=0.27.2",
    "pyyaml~=6.0.1",
    "django-environ~=0.11.2",
    "django-extensions~=3.2.3",
    "django-debug-toolbar~=4.3.0",
    "pre-commit~=3.7.0",
    "ruff==0.3.7",
    "isort==5.13.2",
    "django-upgrade==1.16.0",
    "psycopg[binary]~=3.1.18",
    "django-unfold==0.26.0",
    "pgvector",
    "auto_dataclass",
]
[tool.pip-tools]
extra-index-url = ["https://download.pytorch.org/whl/cu118"]

[tool.isort]
profile = "black"

[tool.ruff]
line-length = 120
